feat(config): export diffProjectConfig from the public entrypoint
Studio is now a second real consumer of the config-diff classification
engine, so promote diffProjectConfig/ConfigChange* from
@supabase/config/internal to the public `.` entrypoint per ADR 0022's
deferred-promotion condition.
