+++ apps/cli/src/legacy/commands/config/diff/diff.handler.ts
@@ -1,9 +1,10 @@
 import {
   CLI_CONFIG_SCHEMA_URL,
+  diffProjectConfig,
   fromApiProjectConfig,
   ProjectConfigParseError,
 } from "@supabase/config";
-import { diffProjectConfig, loadCliConfig } from "@supabase/config/internal";
+import { loadCliConfig } from "@supabase/config/internal";
 import { operationDefinitions } from "@supabase/api/effect";
 import { Effect, Option } from "effect";
