+++ packages/config/src/index.ts
@@ -66,3 +66,11 @@ export {
   unmappedApiFields,
 } from "./project-config/project-config.ts";
 export { ProjectConfigSchema, toProjectConfigJsonSchema } from "./project-config/project-schema.ts";
+export {
+  type ConfigChange,
+  type ConfigChangeClass,
+  type ConfigChangeCounts,
+  type ConfigChangeSet,
+  type DiffProjectConfigOptions,
+  diffProjectConfig,
+} from "./config-diff.ts";
